+++ packages/components/molecules/f-restaurant-card/CHANGELOG.md
@@ -14,6 +14,7 @@ v0.9.0
 ### Changed
 - Updated storybook for restaurant card
 - Renamed RestaurantBadge and RestaurantBadges to RestaurantTag and RestaurantTags to align with language across the business
+- Remove pie token import from tag component and use css class for default colours
 
 v0.8.0
 ------------------------------